feat(frontend): redis、tendbcluster的webcosole #5905 # Reviewed, transaction id: 15707
